Summarise Bootstrap traits

Summarizes the mean and confidence interval for each trait moment.

trait_summarise_boot_moments(
  bootstrap_moments,
  parametric = TRUE,
  sd_mult = 1,
  ci = 0.95
)

Arguments

bootstrap_moments
trait moments from trait_np_bootstrap or trait_parametric_bootstrap

parametric
logical; default is TRUE. Should Confidence Intervals be calculated parametrically (using the mean and SD) or nonparametrically (using quantiles).

sd_mult
Number of standard deviations around each moment, defaults to one

ci
Desired confidence level for use when parametric is false. Defaults to 0.95.

Value

tibble with the grouping variables and the mean of each moment (+/- sd_mult * SD)

Examples

data(community)
data(trait)
imputed_traits <-trait_impute(comm = community, traits = trait,
                 scale_hierarchy = c("Site", "PlotID"),
                 taxon_col = "Taxon", value_col = "Value",
                 trait_col = "Trait", abundance_col = "Cover")
boot_traits <- trait_np_bootstrap(imputed_traits)
trait_summarise_boot_moments(boot_traits)

#> # A tibble: 28 x 17
#> # Groups:   global, Site, PlotID [14]
#>    global Site  PlotID Trait     n    mean ci_low_mean ci_high_mean     var
#>    <chr>  <chr> <chr>  <chr> <int>   <dbl>       <dbl>        <dbl>   <dbl>
#>  1 global 1     A      Plan…   100 2.49        2.28          2.69   8.57e+0
#>  2 global 1     A      Wet_…   100 0.0128      0.0122        0.0134 6.81e-5
#>  3 global 1     B      Plan…   100 1.93        1.81          2.05   3.15e+0
#>  4 global 1     B      Wet_…   100 0.0127      0.0122        0.0132 4.51e-5
#>  5 global 1     C      Plan…   100 1.75        1.67          1.84   1.85e+0
#>  6 global 1     C      Wet_…   100 0.0137      0.0131        0.0144 1.26e-4
#>  7 global 1     D      Plan…   100 1.46        1.34          1.58   3.08e+0
#>  8 global 1     D      Wet_…   100 0.00963     0.00920       0.0101 3.71e-5
#>  9 global 1     E      Plan…   100 7.01        6.73          7.30   1.71e+1
#> 10 global 1     E      Wet_…   100 0.0156      0.0145        0.0168 2.76e-4
#> # … with 18 more rows, and 8 more variables: ci_low_var <dbl>,
#> #   ci_high_var <dbl>, skew <dbl>, ci_low_skew <dbl>, ci_high_skew <dbl>,
#> #   kurt <dbl>, ci_low_kurt <dbl>, ci_high_kurt <dbl>