Summarise Bootstrap traits

Summarizes the mean and confidence interval for each trait moment.

trait_summarise_boot_moments(
  bootstrap_moments,
  parametric = TRUE,
  sd_mult = 1,
  ci = 0.95
)

Arguments

bootstrap_moments

trait moments from trait_np_bootstrap or trait_parametric_bootstrap

parametric

logical; default is TRUE. Should Confidence Intervals be calculated parametrically (using the mean and SD) or nonparametrically (using quantiles).

sd_mult

Number of standard deviations around each moment, defaults to one

ci

Desired confidence level for use when parametric is false. Defaults to 0.95.

Value

tibble with the grouping variables and the mean of each moment (+/- sd_mult * SD)

Examples

data(community)
data(trait)
filled_traits <- trait_fill(
  comm = community, traits = trait,
  scale_hierarchy = c("Site", "PlotID"),
  taxon_col = "Taxon", value_col = "Value",
  trait_col = "Trait", abundance_col = "Cover"
)
boot_traits <- trait_np_bootstrap(filled_traits)
trait_summarise_boot_moments(boot_traits)
#> # A tibble: 42 × 17
#> # Groups:   global, Site, PlotID [14]
#>    global Site  PlotID Trait           n   mean ci_low_mean ci_high_mean     var
#>    <chr>  <chr> <chr>  <chr>       <int>  <dbl>       <dbl>        <dbl>   <dbl>
#>  1 global 1     A      Leaf_Thick…   100 0.214       0.209        0.218  3.49e-3
#>  2 global 1     A      Plant_Heig…   100 1.62        1.50         1.75   3.31e+0
#>  3 global 1     A      Wet_Mass_g    100 0.0109      0.0104       0.0113 4.83e-5
#>  4 global 1     B      Leaf_Thick…   100 0.203       0.200        0.206  1.68e-3
#>  5 global 1     B      Plant_Heig…   100 1.62        1.49         1.76   3.19e+0
#>  6 global 1     B      Wet_Mass_g    100 0.0122      0.0118       0.0127 4.65e-5
#>  7 global 1     C      Leaf_Thick…   100 0.221       0.218        0.224  2.82e-3
#>  8 global 1     C      Plant_Heig…   100 1.72        1.63         1.81   1.58e+0
#>  9 global 1     C      Wet_Mass_g    100 0.0159      0.0152       0.0167 1.12e-4
#> 10 global 1     D      Leaf_Thick…   100 0.201       0.198        0.204  1.77e-3
#> # … with 32 more rows, and 8 more variables: ci_low_var <dbl>,
#> #   ci_high_var <dbl>, skew <dbl>, ci_low_skew <dbl>, ci_high_skew <dbl>,
#> #   kurt <dbl>, ci_low_kurt <dbl>, ci_high_kurt <dbl>